chatapp/requirements.txt
asgiref==3.8.1
channels==4.1.0
Django== 4.2.14
django-cors-headers==4.4.0
djangorestframework==3.15.2
gunicorn==22.0.0
packaging==24.1
sqlparse==0.5.1
